Refactor: Move filter by file type to new file
