+++ src/main.rs
@@ -10,7 +10,8 @@ use self::display::draw_it;
 use clap::{crate_version, Arg};
 use clap::{Command, Values};
 use dir_walker::{walk_it, WalkData};
-use filter::{get_all_file_types, get_biggest};
+use filter::get_biggest;
+use filter_type::get_all_file_types;
 use regex::Regex;
 use std::cmp::max;
 use std::path::PathBuf;
@@ -22,6 +23,7 @@ mod dir_walker;
 mod display;
 mod display_node;
 mod filter;
+mod filter_type;
 mod node;
 mod platform;
 mod utils;
